+++ models/topics.model.js
@@ -1,0 +1,8 @@
+const db = require("../db/connection");
+
+exports.selectTopics = () => {
+    const sqlString = `
+        SELECT * FROM topics;
+    `;
+    return db.query(sqlString);
+};
