+++ models/users.model.js
@@ -1,0 +1,10 @@
+const db = require("../db/connection");
+
+exports.selectUsers = () => {
+
+    const sqlString = `
+        SELECT * FROM users;
+    `
+    return db.query(sqlString);
+
+};
